minor change on Makefile

+++ Makefile
@@ -68,12 +68,10 @@ install: ## Make a conda env with dist/*.whl and dist/*.tar.gz installed
 	conda env remove -y -n $(ENV)-install
 
 bdist: ## Make a dist/*.whl binary distribution
-	#$(SA) $(ENV) && python setup.py bdist_wheel $(POST_SDIST) -r pypi && rm -rf *.egg-info
-	python setup.py bdist_wheel $(POST_SDIST) -r pypi && rm -rf *.egg-info
+	$(SA) $(ENV) && python setup.py bdist_wheel $(POST_SDIST) -r pypi && rm -rf *.egg-info
 
 sdist: ## Make a dist/*.tar.gz source distribution
-	# $(SA) $(ENV) && python setup.py sdist $(POST_SDIST) -r pypi && rm -rf *.egg-info
-	 python setup.py sdist $(POST_SDIST) -r pypi #&& rm -rf *.egg-info
+	$(SA) $(ENV) && python setup.py sdist $(POST_SDIST) -r pypi && rm -rf *.egg-info
 
 test: TEST?=
 test: ## Make a python3 test run
